Create sascodediabetes.sas
SAS code for descriptives of diabetic and non-diabetic factors and MANOVA test to test differences in mean vectors. Ran ANOVA F-tests to determine which variables contribute significantly to the differences in mean vectors. To determine where the differences lie,  ran a tukey test on each variable that produced a significant result.